# GH-94822: Respect metaclasses when caching class attributes