This patch changes `LOAD_ATTR_CLASS` to:
- fail if the named attribute is present on the metaclass
- fail if the metaclass defines `__getattribute__`
- use the metaclass type version for additional cache validation

It also adds a bunch of regression tests for nasty edge-cases when caching class attributes and methods.

<!-- gh-issue-number: gh-94822 -->
* Issue: gh-94822
<!-- /gh-issue-number -->