@brandtbucher 2022-07-14
Note to self: this might be able to be automatically backported by the bots. We just need to change all instances of the term `ATTR` to `METHOD` in the resulting patch.