@brandtbucher 2022-07-14
Looks like this patch makes things about 1% slower. It's not immediately clear to me if this is mainly from removing illegitimate specializations, the increased cost of specializing class attribute loads, or the overhead of additional metaclass version checks:

<details>

```
Slower (33):
- scimark_sparse_mat_mult: 3.93 ms +- 0.09 ms -> 4.13 ms +- 0.05 ms: 1.05x slower
- json_dumps: 12.0 ms +- 0.1 ms -> 12.5 ms +- 0.3 ms: 1.03x slower
- xml_etree_iterparse: 103 ms +- 2 ms -> 107 ms +- 2 ms: 1.03x slower
- logging_silent: 88.9 ns +- 1.7 ns -> 91.5 ns +- 0.6 ns: 1.03x slower
- spectral_norm: 93.2 ms +- 1.6 ms -> 95.5 ms +- 2.3 ms: 1.02x slower
- richards: 44.5 ms +- 0.9 ms -> 45.5 ms +- 1.0 ms: 1.02x slower
- scimark_fft: 303 ms +- 5 ms -> 310 ms +- 3 ms: 1.02x slower
- pickle_pure_python: 287 us +- 3 us -> 293 us +- 3 us: 1.02x slower
- chaos: 65.6 ms +- 1.0 ms -> 67.0 ms +- 1.0 ms: 1.02x slower
- thrift: 728 us +- 14 us -> 740 us +- 12 us: 1.02x slower
- pathlib: 17.8 ms +- 0.2 ms -> 18.1 ms +- 0.3 ms: 1.02x slower
- deltablue: 3.21 ms +- 0.04 ms -> 3.25 ms +- 0.04 ms: 1.01x slower
- regex_compile: 126 ms +- 1 ms -> 128 ms +- 1 ms: 1.01x slower
- scimark_lu: 103 ms +- 2 ms -> 105 ms +- 2 ms: 1.01x slower
- go: 132 ms +- 1 ms -> 133 ms +- 1 ms: 1.01x slower
- pyflate: 393 ms +- 4 ms -> 398 ms +- 5 ms: 1.01x slower
- json: 4.64 ms +- 0.12 ms -> 4.69 ms +- 0.11 ms: 1.01x slower
- logging_format: 6.27 us +- 0.11 us -> 6.34 us +- 0.11 us: 1.01x slower
- nqueens: 78.6 ms +- 0.9 ms -> 79.5 ms +- 1.2 ms: 1.01x slower
- hexiom: 5.92 ms +- 0.02 ms -> 5.98 ms +- 0.02 ms: 1.01x slower
- tornado_http: 91.5 ms +- 1.3 ms -> 92.4 ms +- 1.3 ms: 1.01x slower
- regex_v8: 21.5 ms +- 0.3 ms -> 21.7 ms +- 0.2 ms: 1.01x slower
- xml_etree_parse: 157 ms +- 3 ms -> 158 ms +- 5 ms: 1.01x slower
- nbody: 91.8 ms +- 1.4 ms -> 92.6 ms +- 0.7 ms: 1.01x slower
- sympy_sum: 158 ms +- 1 ms -> 159 ms +- 2 ms: 1.01x slower
- telco: 6.41 ms +- 0.09 ms -> 6.46 ms +- 0.09 ms: 1.01x slower
- unpickle_list: 4.98 us +- 0.04 us -> 5.01 us +- 0.06 us: 1.01x slower
- 2to3: 248 ms +- 1 ms -> 250 ms +- 1 ms: 1.01x slower
- scimark_sor: 110 ms +- 1 ms -> 110 ms +- 1 ms: 1.01x slower
- crypto_pyaes: 72.2 ms +- 0.5 ms -> 72.5 ms +- 0.5 ms: 1.00x slower
- python_startup: 8.19 ms +- 0.01 ms -> 8.22 ms +- 0.01 ms: 1.00x slower
- python_startup_no_site: 6.05 ms +- 0.01 ms -> 6.06 ms +- 0.00 ms: 1.00x slower
- flaskblogging: 4.26 ms +- 0.01 ms -> 4.26 ms +- 0.02 ms: 1.00x slower

Faster (12):
- pickle_list: 4.14 us +- 0.04 us -> 3.98 us +- 0.05 us: 1.04x faster
- raytrace: 294 ms +- 7 ms -> 286 ms +- 6 ms: 1.03x faster
- regex_dna: 207 ms +- 0 ms -> 202 ms +- 0 ms: 1.02x faster
- regex_effbot: 3.58 ms +- 0.03 ms -> 3.51 ms +- 0.01 ms: 1.02x faster
- mako: 9.47 ms +- 0.03 ms -> 9.30 ms +- 0.06 ms: 1.02x faster
- pickle: 10.4 us +- 0.1 us -> 10.2 us +- 0.1 us: 1.02x faster
- pickle_dict: 30.6 us +- 0.1 us -> 30.3 us +- 0.1 us: 1.01x faster
- chameleon: 6.43 ms +- 0.15 ms -> 6.38 ms +- 0.05 ms: 1.01x faster
- scimark_monte_carlo: 65.6 ms +- 0.9 ms -> 65.1 ms +- 1.1 ms: 1.01x faster
- django_template: 32.2 ms +- 0.4 ms -> 31.9 ms +- 0.3 ms: 1.01x faster
- json_loads: 24.3 us +- 0.2 us -> 24.2 us +- 0.2 us: 1.01x faster
- xml_etree_process: 52.2 ms +- 0.6 ms -> 52.0 ms +- 0.5 ms: 1.00x faster

Benchmark hidden because not significant (16): dulwich_log, fannkuch, float, html5lib, logging_simple, meteor_contest, pidigits, pycparser, sqlite_synth, sympy_expand, sympy_integrate, sympy_str, unpack_sequence, unpickle, unpickle_pure_python, xml_etree_generate

Geometric mean: 1.01x slower
```

</details>

Total hit rates for `LOAD_ATTR_CLASS` have gone down by 15%, but hit rates for the whole `LOAD_ATTR` family are largely unchanged. Specialization successes decreased by 3%. 5% more failures are attributed to metaclass `__getattribute__` methods, while 7% more failures are due to the named attributes being present on the metaclass itself.

## `main`

|Name | Count | Self | Cumulative | Miss ratio | 
|---|---:|---:|---:|---:|
| LOAD_ATTR_CLASS | 102,062,535 | 0.2% | 96.5% | 1.6% |

<details>

<summary> specialization stats for LOAD_ATTR family </summary>

|Kind | Count | Ratio | 
|---|---|---|
|  unquickened |    193618648 | 3.4% |
| specialization.deferred |    629431714 | 11.2% |
| specialization.deopt |      2359119 | 0.0% |
|          hit |   4675648068 | 83.1% |
|         miss |    129400942 | 2.3% |

#### Specialization attempts

| | Count | Ratio | 
|---|---:|---:|
| Success | 4,958,258 | 64.8% |
| Failure | 2,693,208 | 35.2% |

|Failure kind | Count | Ratio | 
|---|---:|---:|
| not managed dict | 586,470 | 21.8% |
| overridden | 540,857 | 20.1% |
| method | 384,095 | 14.3% |
| non object slot | 264,897 | 9.8% |
| out of range | 254,521 | 9.5% |
| metaclass attribute | 187,132 | 6.9% |
| non overriding descriptor | 171,346 | 6.4% |
| class method obj | 152,554 | 5.7% |
| has managed dict | 116,971 | 4.3% |
| mutable class | 112,244 | 4.2% |
| builtin class method | 53,527 | 2.0% |
| shadowed | 3,840 | 0.1% |
| other | 1,669 | 0.1% |
| module attr not found | 1,648 | 0.1% |
| expected error | 591 | 0.0% |
| property | 401 | 0.0% |

</details>

## This PR

|Name | Count | Self | Cumulative | Miss ratio | 
|---|---:|---:|---:|---:|
| LOAD_ATTR_CLASS | 88,312,319 | 0.1% | 97.0% | 0.4% |

<details>

<summary> specialization stats for LOAD_ATTR family </summary>

|Kind | Count | Ratio | 
|---|---|---|
|  unquickened |    190837306 | 3.4% |
| specialization.deferred |    636947625 | 11.3% |
| specialization.deopt |      2321544 | 0.0% |
|          hit |   4660987261 | 83.0% |
|         miss |    126561331 | 2.3% |

#### Specialization attempts

| | Count | Ratio | 
|---|---:|---:|
| Success | 4,829,110 | 61.4% |
| Failure | 3,032,191 | 38.6% |

|Failure kind | Count | Ratio | 
|---|---:|---:|
| overridden | 765,516 | 25.2% |
| not managed dict | 582,142 | 19.2% |
| metaclass attribute | 417,214 | 13.8% |
| method | 383,456 | 12.6% |
| non object slot | 265,664 | 8.8% |
| out of range | 253,965 | 8.4% |
| class method obj | 144,317 | 4.8% |
| has managed dict | 117,040 | 3.9% |
| mutable class | 108,085 | 3.6% |
| non overriding descriptor | 74,494 | 2.5% |
| builtin class method | 53,548 | 1.8% |
| shadowed | 3,840 | 0.1% |
| module attr not found | 1,648 | 0.1% |
| expected error | 591 | 0.0% |
| property | 380 | 0.0% |

</details>