@brandtbucher 2022-07-14
I might prepare a second branch that disallows metaclasses entirely, and compare the results.