@Fidget-Spinner 2022-07-15
Wow, these are really comprehensive stats. Thanks Brandt.

FYI, enum attribute access isn't specialised in 3.11. So whatever we do it probably won't hurt 3.11 that much :).