@brandtbucher 2022-07-15
Disallowing metaclasses entirely (https://github.com/python/cpython/compare/main...brandtbucher:no-metaclasses) gets rid of the 1% slowdown for the benchmark suite as a whole. It looks like it does come with a ~1% hit to the suite's `sympy` benchmarks when compared to this PR, but that's actually much better than I was expecting:

<details>

```
Slower (27):
- nqueens: 78.6 ms +- 0.9 ms -> 82.1 ms +- 0.9 ms: 1.04x slower
- fannkuch: 370 ms +- 5 ms -> 386 ms +- 7 ms: 1.04x slower
- go: 132 ms +- 1 ms -> 137 ms +- 1 ms: 1.04x slower
- sqlite_synth: 2.59 us +- 0.05 us -> 2.64 us +- 0.07 us: 1.02x slower
- thrift: 728 us +- 14 us -> 743 us +- 20 us: 1.02x slower
- regex_compile: 126 ms +- 1 ms -> 129 ms +- 1 ms: 1.02x slower
- sympy_sum: 158 ms +- 1 ms -> 160 ms +- 2 ms: 1.02x slower
- pidigits: 190 ms +- 0 ms -> 193 ms +- 0 ms: 1.02x slower
- logging_silent: 88.9 ns +- 1.7 ns -> 90.2 ns +- 1.7 ns: 1.01x slower
- regex_v8: 21.5 ms +- 0.3 ms -> 21.8 ms +- 0.2 ms: 1.01x slower
- hexiom: 5.92 ms +- 0.02 ms -> 5.98 ms +- 0.03 ms: 1.01x slower
- pathlib: 17.8 ms +- 0.2 ms -> 18.0 ms +- 0.2 ms: 1.01x slower
- pickle_dict: 30.6 us +- 0.1 us -> 31.0 us +- 0.1 us: 1.01x slower
- scimark_lu: 103 ms +- 2 ms -> 104 ms +- 2 ms: 1.01x slower
- dulwich_log: 61.1 ms +- 0.4 ms -> 61.7 ms +- 0.4 ms: 1.01x slower
- pyflate: 393 ms +- 4 ms -> 397 ms +- 5 ms: 1.01x slower
- sympy_str: 279 ms +- 4 ms -> 281 ms +- 6 ms: 1.01x slower
- sympy_integrate: 20.1 ms +- 0.1 ms -> 20.3 ms +- 0.2 ms: 1.01x slower
- unpickle_list: 4.98 us +- 0.04 us -> 5.02 us +- 0.09 us: 1.01x slower
- pickle_list: 4.14 us +- 0.04 us -> 4.17 us +- 0.03 us: 1.01x slower
- 2to3: 248 ms +- 1 ms -> 250 ms +- 1 ms: 1.01x slower
- crypto_pyaes: 72.2 ms +- 0.5 ms -> 72.8 ms +- 0.7 ms: 1.01x slower
- xml_etree_iterparse: 103 ms +- 2 ms -> 104 ms +- 2 ms: 1.01x slower
- python_startup: 8.19 ms +- 0.01 ms -> 8.24 ms +- 0.01 ms: 1.01x slower
- python_startup_no_site: 6.05 ms +- 0.01 ms -> 6.08 ms +- 0.01 ms: 1.01x slower
- scimark_fft: 303 ms +- 5 ms -> 304 ms +- 3 ms: 1.00x slower
- flaskblogging: 4.26 ms +- 0.01 ms -> 4.26 ms +- 0.01 ms: 1.00x slower

Faster (20):
- regex_effbot: 3.58 ms +- 0.03 ms -> 3.37 ms +- 0.01 ms: 1.06x faster
- raytrace: 294 ms +- 7 ms -> 282 ms +- 4 ms: 1.04x faster
- regex_dna: 207 ms +- 0 ms -> 202 ms +- 1 ms: 1.02x faster
- scimark_monte_carlo: 65.6 ms +- 0.9 ms -> 64.4 ms +- 0.5 ms: 1.02x faster
- scimark_sparse_mat_mult: 3.93 ms +- 0.09 ms -> 3.86 ms +- 0.08 ms: 1.02x faster
- nbody: 91.8 ms +- 1.4 ms -> 90.3 ms +- 0.5 ms: 1.02x faster
- pickle: 10.4 us +- 0.1 us -> 10.2 us +- 0.1 us: 1.02x faster
- spectral_norm: 93.2 ms +- 1.6 ms -> 91.8 ms +- 0.5 ms: 1.01x faster
- logging_format: 6.27 us +- 0.11 us -> 6.19 us +- 0.07 us: 1.01x faster
- logging_simple: 5.69 us +- 0.08 us -> 5.63 us +- 0.07 us: 1.01x faster
- chameleon: 6.43 ms +- 0.15 ms -> 6.36 ms +- 0.05 ms: 1.01x faster
- django_template: 32.2 ms +- 0.4 ms -> 31.9 ms +- 0.3 ms: 1.01x faster
- json_dumps: 12.0 ms +- 0.1 ms -> 11.9 ms +- 0.1 ms: 1.01x faster
- xml_etree_process: 52.2 ms +- 0.6 ms -> 51.7 ms +- 0.4 ms: 1.01x faster
- deltablue: 3.21 ms +- 0.04 ms -> 3.18 ms +- 0.05 ms: 1.01x faster
- json_loads: 24.3 us +- 0.2 us -> 24.1 us +- 0.2 us: 1.01x faster
- chaos: 65.6 ms +- 1.0 ms -> 65.2 ms +- 0.8 ms: 1.01x faster
- xml_etree_generate: 75.0 ms +- 0.7 ms -> 74.6 ms +- 0.4 ms: 1.01x faster
- scimark_sor: 110 ms +- 1 ms -> 109 ms +- 1 ms: 1.00x faster
- mako: 9.47 ms +- 0.03 ms -> 9.44 ms +- 0.04 ms: 1.00x faster

Benchmark hidden because not significant (14): float, html5lib, json, meteor_contest, pickle_pure_python, pycparser, richards, sympy_expand, telco, tornado_http, unpack_sequence, unpickle, unpickle_pure_python, xml_etree_parse

Geometric mean: 1.00x slower
```

</details>

More stats:

|Name | Count | Self | Cumulative | Miss ratio | 
|---|---:|---:|---:|---:|
| LOAD_ATTR_CLASS | 67,240,997 | 0.1% | 97.5% | 0.2% |

<details>

<summary> specialization stats for LOAD_ATTR family </summary>

|Kind | Count | Ratio | 
|---|---|---|
|  unquickened |    192112856 | 3.5% |
| specialization.deferred |    663275886 | 12.0% |
| specialization.deopt |      2346873 | 0.0% |
|          hit |   4553072619 | 82.2% |
|         miss |    127895746 | 2.3% |

#### Specialization attempts

| | Count | Ratio | 
|---|---:|---:|
| Success | 4,842,590 | 61.2% |
| Failure | 3,070,807 | 38.8% |

|Failure kind | Count | Ratio | 
|---|---:|---:|
| metaclass attribute | 724,696 | 23.6% |
| not managed dict | 586,451 | 19.1% |
| overridden | 528,290 | 17.2% |
| method | 383,637 | 12.5% |
| non object slot | 264,709 | 8.6% |
| out of range | 254,092 | 8.3% |
| class method obj | 128,659 | 4.2% |
| has managed dict | 116,914 | 3.8% |
| mutable class | 97,596 | 3.2% |
| non overriding descriptor | 65,531 | 2.1% |
| builtin class method | 53,397 | 1.7% |
| shadowed | 3,840 | 0.1% |
| module attr not found | 1,648 | 0.1% |
| expected error | 591 | 0.0% |
| property | 254 | 0.0% |

</details>

It seems that your earlier intuition was spot-on @Fidget-Spinner. At this point, I'm thinking that we:
- Use my other branch, and fail to specialize in the presence of a metaclass. This will be backported to 3.11.
- Try adding a `LOAD_ATTR_METACLASS` opcode to boost performance for things like `enum` and `sympy` in 3.12. That way we only need to bother with (apparently expensive) metaclass versions in those rare cases.

I'm planning to move forward with these today.